Model_final/requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.2.0
anyio==4.0.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astor==0.8.1
asttokens==2.4.0
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.2
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.4
bokeh==3.2.2
brotlipy==0.7.0
cachetools==4.2.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.4
click==8.0.4
colorama==0.4.6
comm==0.1.4
contourpy==1.1.1
cryptography==38.0.4
cycler==0.11.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
editdistance==0.6.2
exceptiongroup==1.1.3
executing==1.2.0
fastjsonschema==2.18.0
flatbuffers==20210226132247
flit_core==3.6.0
fonttools==4.42.1
fqdn==1.5.1
frozenlist==1.3.3
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.42.0
h5py==3.7.0
idna==3.4
importlib-metadata==4.11.3
importlib-resources==6.0.1
ipykernel==6.25.2
ipython==8.15.0
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter_client==8.3.1
jupyter-console==6.6.3
jupyter_core==5.3.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.0
jupyterlab-widgets==3.0.9
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
livelossplot==0.5.5
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.8.0
matplotlib-inline==0.1.6
mistune==3.0.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
multidict==6.0.2
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.8
notebook==7.0.3
notebook_shim==0.2.3
numpy==1.23.5
oauthlib==3.2.1
opencv-python==4.8.0.76
opt-einsum==3.3.0
overrides==7.4.0
packaging==22.0
pandas==2.1.0
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.1
pip==22.3.1
platformdirs==3.10.0
pooch==1.4.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==3.20.1
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.16.1
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.11
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.3
rsa==4.7.2
scikit-learn==1.3.0
scipy==1.10.0
Send2Trash==1.8.2
setuptools==65.6.3
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.2
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==2.1.0
terminado==0.17.1
threadpoolctl==3.2.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.3.3
traitlets==5.10.0
typing_extensions==4.4.0
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.14
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
Werkzeug==2.2.2
wheel==0.35.1
widgetsnbextension==4.0.9
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.14.1
xyzservices==2023.7.0
yarl==1.8.1
zipp==3.11.0